termux_article_cli/requirements.txt
python-dotenv
openai
huggingface_hub
google-generativeai
PyYAML
